% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewrapper.R
\name{eWrapper}
\alias{eWrapper}
\title{eWrapper closure for message decoding}
\usage{
eWrapper()
}
\value{
an eWrapper closure
}
\description{
eWrapper closure for message decoding
}
\examples{
Examples here
}
