% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ibrokers2-package.R
\docType{package}
\name{Ibrokers2-package}
\alias{Ibrokers2}
\alias{Ibrokers2-package}
\title{Ibrokers2: R API to Interactive Brokers}
\description{
Provides a native R access to Interactive Brokers via the Interactive Brokers API.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
