% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.tws_con}
\alias{print.tws_con}
\title{Printing tws connection}
\usage{
\method{print}{tws_con}(x, ...)
}
\arguments{
\item{x}{object to print}

\item{...}{Other arguments passed on to individual methods.}
}
\description{
Printing tws connection
}
\examples{
print(tws_con)
}
