% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_current_time.R
\name{req_current_time}
\alias{req_current_time}
\title{Request the current time from the IB server}
\usage{
req_current_time(tws_con)
}
\arguments{
\item{tws_con}{a valid tws connection.}
}
\value{
returns TWS's current time.
}
\description{
TWS's current time. TWS is synchronized with the server (not local computer) 
using NTP and this function will receive the current time in TWS.
}
\examples{
example here
}
