% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqids.R
\name{reqIds}
\alias{reqIds}
\title{Request next valid Id}
\usage{
reqIds(con, numIds = 1L)
}
\arguments{
\item{con}{}

\item{numIds}{}
}
\value{
returns the next valid Id that can be used as an order id.
}
\description{
Request next valid Id
}
\examples{
example here
}
