% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tws_connect.R
\name{tws_connect}
\alias{tws_connect}
\alias{is_tws_connection}
\alias{tws_disconnect}
\alias{is_tws_connected}
\title{Establish a connection to TWS}
\usage{
tws_connect(clientId = 1L, host = "127.0.0.1", port = 7497,
  verbose = TRUE, timeout = 5, blocking = .Platform$OS.type ==
  "windows")

is_tws_connection(x)

tws_disconnect(tws_con)

is_tws_connected(tws_con)
}
\arguments{
\item{clientId}{}

\item{host}{Default connection is 127.0.0.1 for the trade work station}

\item{port}{Default connection is to the paper trading account on port 7497.
Live connection to tws is on port 7496. For the IB gateway the live port is 4001 
and the paper trading port is 4002.}

\item{verbose}{}

\item{timeout}{}

\item{blocking}{}
}
\value{
A tws_con object.
}
\description{
Establish a connection to TWS
}
\examples{
\dontrun{
tws <- tws_connect()
tws_disconnect(tws)
}
}
