% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_messages.R
\name{process_messages}
\alias{process_messages}
\title{Process messages from TWS}
\usage{
process_messages(curMsg, tws_con, ...)
}
\arguments{
\item{curMsg}{}

\item{tws_con}{a valid tws connection object}

\item{...}{}
}
\value{
used for processing incoming messages. Called for it's side-effects.
}
\description{
Process messages from TWS
}
\examples{
examples here
}
\keyword{internal}
