% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_current_time.R
\name{req_current_time}
\alias{req_current_time}
\title{Request the current time from the IB server}
\usage{
req_current_time(con)
}
\arguments{
\item{con}{}
}
\value{
returns the next valid Id that can be used as an order id.
}
\description{
Request the current time from the IB server
}
\examples{
example here
}
