% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqids.R
\name{reqIds}
\alias{reqIds}
\title{Request next valid Id}
\usage{
reqIds(tws_con)
}
\arguments{
\item{tws_con}{a valid tws_con connection object.}
}
\value{
returns the next valid Id.
}
\description{
Requests the next valid order ID at the current moment.
}
\examples{
\dontrun{
tws_con <- tws_connect()
reqIds(tws_con)
}
}
