% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tws_connect.R
\name{tws_connect}
\alias{tws_connect}
\alias{is_tws_connection}
\alias{tws_disconnect}
\alias{is_tws_connected}
\title{Establish a connection to TWS}
\usage{
tws_connect(clientId = 1L, host = "localhost", port = 7496,
  verbose = TRUE, timeout = 5, filename = NULL,
  blocking = .Platform$OS.type == "windows")

is_tws_connection(x)

tws_disconnect(tws_con)

is_tws_connected(tws_con)
}
\arguments{
\item{clientId}{}

\item{host}{}

\item{port}{}

\item{verbose}{}

\item{timeout}{}

\item{filename}{}

\item{blocking}{}
}
\value{
A tws_con object.
}
\description{
Establish a connection to TWS
}
\examples{
\dontrun{
tws <- tws_connect()
tws_disconnect(tws)
}
}
